Fixing Duplication Error
